requirements.txt
faiss-cpu
h5py
numpy
requests
matplotlib
diskann-pybind